spellchecking and base loader
